Make axisartist.axis_artist pseudo-tests real.
The results seem to be acceptable.
